requirements.txt
# until I have a tool to prevent this, this requirements.txt mixes happily dev and prod dependencies

appnope==0.1.0
asn1crypto==0.24.0
backcall==0.1.0
cffi==1.11.5
cryptography==2.3.1
decorator==4.3.0
Django==2.1.1
django-rest-knox==3.1.5
djangorestframework==3.8.2
idna==2.7
ipdb==0.11
ipython==7.0.1
ipython-genutils==0.2.0
jedi==0.12.1
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.5
psycopg2==2.7.5
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.2.0
pyOpenSSL==18.0.0
pytz==2018.5
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
wcwidth==0.1.7
